server json is now sending. WIP for getting client rendering.
