+++ api/authMethod.js
@@ -1,0 +1,4 @@
+module.exports = (request, response) => {
+  const data = {'name': 'person'}
+  response.json(data)
+}
