+++ client/index.html
@@ -16,7 +16,7 @@
   </head>
   <body>
     <h1>Helllo</h1>
-    <main></main>
-    <script src='./build/bundle.js'></script>
+    <div id="DontCall"></div>
+    <script src='/bundle.js'></script>
   </body>
 </html>
